Move waze_travel_time services to separate module
