+++ homeassistant/components/waze_travel_time/services.py
@@ -1,0 +1,243 @@
+"""The waze_travel_time component."""
+
+from collections.abc import Collection
+import logging
+from typing import Literal
+
+from pywaze.route_calculator import CalcRoutesResponse, WazeRouteCalculator, WRCError
+import voluptuous as vol
+
+from homeassistant.const import CONF_REGION, UnitOfLength
+from homeassistant.core import (
+    HomeAssistant,
+    ServiceCall,
+    ServiceResponse,
+    SupportsResponse,
+)
+from homeassistant.helpers.httpx_client import get_async_client
+from homeassistant.helpers.location import find_coordinates
+from homeassistant.helpers.selector import (
+    BooleanSelector,
+    SelectSelector,
+    SelectSelectorConfig,
+    SelectSelectorMode,
+    TextSelector,
+    TextSelectorConfig,
+    TextSelectorType,
+)
+from homeassistant.util.unit_conversion import DistanceConverter
+
+from .const import (
+    CONF_AVOID_FERRIES,
+    CONF_AVOID_SUBSCRIPTION_ROADS,
+    CONF_AVOID_TOLL_ROADS,
+    CONF_DESTINATION,
+    CONF_EXCL_FILTER,
+    CONF_INCL_FILTER,
+    CONF_ORIGIN,
+    CONF_REALTIME,
+    CONF_UNITS,
+    CONF_VEHICLE_TYPE,
+    DEFAULT_FILTER,
+    DEFAULT_VEHICLE_TYPE,
+    DOMAIN,
+    IMPERIAL_UNITS,
+    METRIC_UNITS,
+    REGIONS,
+    UNITS,
+    VEHICLE_TYPES,
+)
+
+SERVICE_GET_TRAVEL_TIMES = "get_travel_times"
+SERVICE_GET_TRAVEL_TIMES_SCHEMA = vol.Schema(
+    {
+        vol.Required(CONF_ORIGIN): TextSelector(),
+        vol.Required(CONF_DESTINATION): TextSelector(),
+        vol.Required(CONF_REGION): SelectSelector(
+            SelectSelectorConfig(
+                options=REGIONS,
+                mode=SelectSelectorMode.DROPDOWN,
+                translation_key=CONF_REGION,
+                sort=True,
+            )
+        ),
+        vol.Optional(CONF_REALTIME, default=False): BooleanSelector(),
+        vol.Optional(CONF_VEHICLE_TYPE, default=DEFAULT_VEHICLE_TYPE): SelectSelector(
+            SelectSelectorConfig(
+                options=VEHICLE_TYPES,
+                mode=SelectSelectorMode.DROPDOWN,
+                translation_key=CONF_VEHICLE_TYPE,
+                sort=True,
+            )
+        ),
+        vol.Optional(CONF_UNITS, default=METRIC_UNITS): SelectSelector(
+            SelectSelectorConfig(
+                options=UNITS,
+                mode=SelectSelectorMode.DROPDOWN,
+                translation_key=CONF_UNITS,
+                sort=True,
+            )
+        ),
+        vol.Optional(CONF_AVOID_TOLL_ROADS, default=False): BooleanSelector(),
+        vol.Optional(CONF_AVOID_SUBSCRIPTION_ROADS, default=False): BooleanSelector(),
+        vol.Optional(CONF_AVOID_FERRIES, default=False): BooleanSelector(),
+        vol.Optional(CONF_INCL_FILTER): TextSelector(
+            TextSelectorConfig(
+                type=TextSelectorType.TEXT,
+                multiple=True,
+            ),
+        ),
+        vol.Optional(CONF_EXCL_FILTER): TextSelector(
+            TextSelectorConfig(
+                type=TextSelectorType.TEXT,
+                multiple=True,
+            ),
+        ),
+    }
+)
+
+_LOGGER = logging.getLogger(__name__)
+
+
+def async_setup_services(hass: HomeAssistant) -> None:
+    """Register integration services."""
+    hass.services.async_register(
+        DOMAIN,
+        SERVICE_GET_TRAVEL_TIMES,
+        _async_get_travel_times_service,
+        SERVICE_GET_TRAVEL_TIMES_SCHEMA,
+        supports_response=SupportsResponse.ONLY,
+    )
+
+
+async def _async_get_travel_times_service(service: ServiceCall) -> ServiceResponse:
+    hass = service.hass
+    httpx_client = get_async_client(hass)
+    client = WazeRouteCalculator(
+        region=service.data[CONF_REGION].upper(), client=httpx_client
+    )
+
+    origin_coordinates = find_coordinates(hass, service.data[CONF_ORIGIN])
+    destination_coordinates = find_coordinates(hass, service.data[CONF_DESTINATION])
+
+    origin = origin_coordinates if origin_coordinates else service.data[CONF_ORIGIN]
+    destination = (
+        destination_coordinates
+        if destination_coordinates
+        else service.data[CONF_DESTINATION]
+    )
+
+    response = await async_get_travel_times(
+        client=client,
+        origin=origin,
+        destination=destination,
+        vehicle_type=service.data[CONF_VEHICLE_TYPE],
+        avoid_toll_roads=service.data[CONF_AVOID_TOLL_ROADS],
+        avoid_subscription_roads=service.data[CONF_AVOID_SUBSCRIPTION_ROADS],
+        avoid_ferries=service.data[CONF_AVOID_FERRIES],
+        realtime=service.data[CONF_REALTIME],
+        units=service.data[CONF_UNITS],
+        incl_filters=service.data.get(CONF_INCL_FILTER, DEFAULT_FILTER),
+        excl_filters=service.data.get(CONF_EXCL_FILTER, DEFAULT_FILTER),
+    )
+    return {"routes": [vars(route) for route in response]} if response else None
+
+
+async def async_get_travel_times(
+    client: WazeRouteCalculator,
+    origin: str,
+    destination: str,
+    vehicle_type: str,
+    avoid_toll_roads: bool,
+    avoid_subscription_roads: bool,
+    avoid_ferries: bool,
+    realtime: bool,
+    units: Literal["metric", "imperial"] = "metric",
+    incl_filters: Collection[str] | None = None,
+    excl_filters: Collection[str] | None = None,
+) -> list[CalcRoutesResponse] | None:
+    """Get all available routes."""
+
+    incl_filters = incl_filters or ()
+    excl_filters = excl_filters or ()
+
+    _LOGGER.debug(
+        "Getting update for origin: %s destination: %s",
+        origin,
+        destination,
+    )
+    routes = []
+    vehicle_type = "" if vehicle_type.upper() == "CAR" else vehicle_type.upper()
+    try:
+        routes = await client.calc_routes(
+            origin,
+            destination,
+            vehicle_type=vehicle_type,
+            avoid_toll_roads=avoid_toll_roads,
+            avoid_subscription_roads=avoid_subscription_roads,
+            avoid_ferries=avoid_ferries,
+            real_time=realtime,
+            alternatives=3,
+        )
+        _LOGGER.debug("Got routes: %s", routes)
+
+        incl_routes: list[CalcRoutesResponse] = []
+
+        def should_include_route(route: CalcRoutesResponse) -> bool:
+            if len(incl_filters) < 1:
+                return True
+            should_include = any(
+                street_name in incl_filters or "" in incl_filters
+                for street_name in route.street_names
+            )
+            if not should_include:
+                _LOGGER.debug(
+                    "Excluding route [%s], because no inclusive filter matched any streetname",
+                    route.name,
+                )
+                return False
+            return True
+
+        incl_routes = [route for route in routes if should_include_route(route)]
+
+        filtered_routes: list[CalcRoutesResponse] = []
+
+        def should_exclude_route(route: CalcRoutesResponse) -> bool:
+            for street_name in route.street_names:
+                for excl_filter in excl_filters:
+                    if excl_filter == street_name:
+                        _LOGGER.debug(
+                            "Excluding route, because exclusive filter [%s] matched streetname: %s",
+                            excl_filter,
+                            route.name,
+                        )
+                        return True
+            return False
+
+        filtered_routes = [
+            route for route in incl_routes if not should_exclude_route(route)
+        ]
+
+        if units == IMPERIAL_UNITS:
+            filtered_routes = [
+                CalcRoutesResponse(
+                    name=route.name,
+                    distance=DistanceConverter.convert(
+                        route.distance, UnitOfLength.KILOMETERS, UnitOfLength.MILES
+                    ),
+                    duration=route.duration,
+                    street_names=route.street_names,
+                )
+                for route in filtered_routes
+                if route.distance is not None
+            ]
+
+        if len(filtered_routes) < 1:
+            _LOGGER.warning("No routes found")
+            return None
+    except WRCError as exp:
+        _LOGGER.warning("Error on retrieving data: %s", exp)
+        return None
+
+    else:
+        return filtered_routes
